create new folder for meeting modal components
